% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{flz_auth}
\alias{flz_auth}
\title{Autenticarse con la API de Fidelizador usando Client Credentials}
\usage{
flz_auth(
  client_id = NULL,
  client_secret = NULL,
  base_url = "https://api.fidelizador.com"
)
}
\arguments{
\item{client_id}{Client ID obtenido del portal de Fidelizador (opcional si ya se sete\if{html}{\out{<c3>}}\if{html}{\out{<b3>}} como variable de entorno)}

\item{client_secret}{Client Secret obtenido del portal de Fidelizador (opcional si ya se sete\if{html}{\out{<c3>}}\if{html}{\out{<b3>}} como variable de entorno)}

\item{base_url}{URL base de la API (por defecto: "https://api.fidelizador.com")}
}
\value{
Token de acceso
}
\description{
Autenticarse con la API de Fidelizador usando Client Credentials
}
