% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{flz_get_credentials}
\alias{flz_get_credentials}
\title{Obtener credenciales desde variables de entorno}
\usage{
flz_get_credentials()
}
\value{
Lista con client_id y client_secret
}
\description{
Obtener credenciales desde variables de entorno
}
