% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{flz_set_credentials}
\alias{flz_set_credentials}
\title{Guardar credenciales en variables de entorno}
\usage{
flz_set_credentials(client_id, client_secret)
}
\arguments{
\item{client_id}{Client ID}

\item{client_secret}{Client Secret}
}
\description{
Guardar credenciales en variables de entorno
}
