% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{flz_get_lists}
\alias{flz_get_lists}
\title{Obtener todas las listas del sistema}
\usage{
flz_get_lists(
  token,
  client_slug,
  format = "json",
  base_url = "https://api.fidelizador.com"
)
}
\arguments{
\item{token}{Token de autenticación obtenido con \code{flz_auth()}}

\item{client_slug}{Slug del cliente de Fidelizador.}

\item{format}{Formato de respuesta: "json" o "xml" (por defecto "json")}

\item{base_url}{URL base de la API (por defecto: "https://api.fidelizador.com")}
}
\value{
Objeto con la respuesta de la API conteniendo todas las listas
}
\description{
Obtiene todas las listas de contactos del sistema en Fidelizador.
}
\examples{
\dontrun{
token <- flz_auth()

# Obtener todas las listas
listas <- flz_get_lists(
  token = token,
  client_slug = "mi-empresa"
)
}
}
